+++ FitCloudKit/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## V1.2.9-beta.182 build20240918001 (2024-09-18)
+    1. Add the `formatedProjNo`, `formatedFirmwareVersion`, `uiVersion`, `screenResolution` and `screenShape` property for FitCloudKitConnectRecord.
+    2. Introduce the API `+(FITCLOUDWATCHCONNECTIONFLAG) connectionFlag;` to indicate whether the current connection is established when the watch device is bound or in the process of reconnecting.
+    
 ## V1.2.9-beta.181 build20240914001 (2024-09-14)
     1. Add the `mainProjNo` and `subProjNo` property for FitCloudPeripheral and FitCloudManufacturer.